requirements.txt
streamlit
gspread
oauth2client
pandas
matplotlib
plotly
streamlit-cookies-manager
google-auth